requirements.txt
scenedetect
ultralytics
torch
torchvision
tqdm
yt-dlp
faster-whisper
openai
python-dotenv
mediapipe==0.10.14
boto3

fastapi
uvicorn
python-multipart
httpx
Pillow
beautifulsoup4
